Report forkpty/exec failures, use _exit over exit.
exit() calls the global destructors, and it does so on the thread that
calls exit().  Destructing the global g_wakeupFd object is a problem,
because other threads (or the SIGWINCH handler) can try to use g_wakeupFd.
Avoid this problem by calling _exit instead.  Improve the fatal function
and use it everywhere.